requirements.txt
python-telegram-bot
google-generativeai
pydub
SpeechRecognition